requirements.txt
packaging
numpy
unsloth_zoo
torchvision
torchaudio
triton
bitsandbytes
peft
transformers==4.57.0
trl==0.22.2